requirements-test.txt
responses~=0.16.0
pytest_testinfra~=6.4.0
testinfra~=6.0.0
pytest~=6.2.5
pytest-cov~=3.0.0
pyyaml
